Moved contact drawer to organisms Staging commit for adding 'add appointment' logic
