+++ package.json
@@ -12,7 +12,7 @@
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
     "@material-ui/pickers": "^3.2.10",
-    "@storybook/addon-knobs": "^6.0.21",
+    "@storybook/addon-knobs": "^6.0.26",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
@@ -67,11 +67,11 @@
   },
   "devDependencies": {
     "@microsoft/microsoft-graph-client": "^2.0.0",
-    "@storybook/addon-actions": "^6.0.21",
-    "@storybook/addon-links": "^6.0.21",
-    "@storybook/addons": "^6.0.21",
+    "@storybook/addon-actions": "^6.0.26",
+    "@storybook/addon-links": "^6.0.26",
+    "@storybook/addons": "^6.0.26",
     "@storybook/preset-create-react-app": "^3.1.4",
-    "@storybook/react": "^6.0.21",
+    "@storybook/react": "^6.0.26",
     "@testing-library/react-hooks": "^3.4.1",
     "@types/bcryptjs": "^2.4.2",
     "@types/lodash": "^4.14.161",
