+++ src/components/organisms/ContactDrawer/ContactDrawer.stories.tsx
@@ -6,17 +6,20 @@ import {GlobalContext} from "../../../common/GlobalContext";
 import {action} from "@storybook/addon-actions";
 
 export default {
-    title: "Atoms/Drawer",
+    title: "Organisms/Drawer",
     component: Drawer,
 };
 
-export const withProps = () => (
+export const Default = () => (
     <GlobalContext.Provider value={{
             saveLibrary: () => action("Calling mock saveLibrary"),
             getAuthenticatedUser: () => action("Calling mock getAuthenticatedUser")
         }}>
         <MockForm>
-            <Drawer library={newLibrary}/>
+            <Drawer
+                library={newLibrary}
+                handleScheduleNextAppointment={action("Submitted")}
+            />
         </MockForm>
     </GlobalContext.Provider>
 );
